+++ raise_utils/data/data.py
@@ -5,7 +5,6 @@ import pandas as pd
 import numpy as np
 from sklearn.model_selection import train_test_split
 from tensorflow.keras.utils import to_categorical
-from raise_utils.hooks import Hook
 
 
 class Data: