+++ raise_utils/hyperparams/__init__.py
@@ -1,3 +1,4 @@
 from raise_utils.hyperparams.dodge import DODGE
-from raise_utils.hyperparams.ghost import GHOST
+from raise_utils.hyperparams.ghost import BinaryGHOST
+from raise_utils.hyperparams.multi_ghost import GHOST
 from raise_utils.hyperparams.raytune import RayTune