+++ raise_utils/utils/__init__.py
@@ -1,1 +1,2 @@
 from raise_utils.utils.data_utils import _check_data
+from raise_utils.utils.utils import *