+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.6.3
+
+* Added G-1 score and IFA (initial false alarm) to metrics (thanks @HuyTu7!)
+* Added `GHOST` and `BinaryGHOST` to `raise_utils.hyperparams`
+
 ## 1.6.2
 
 * Bug fix in `DODGE`